Changed a lot of structure in classes.
Moved DataRetriever from DataRetriever package. DataRetriever is useful to the Worker class. There is a better data correction method.

Changed the types in Measurement so that parsing doesn't have to happen literally everywhere. (It still has to happen in a lot of places but that just can't be helped unfortunately).

TestMain remains as a quick testing thing and I am too lazy to figure out how tests and stuff work. Should not make it into production.